feat({header,info,music}.ily): split header info from music.ily
In an effort to prune the `music.ily` file, which has too much going on
within it, we've started by taking the header information and splitting
it into two stages.

The first stage is the `info.ily` file where we actually specify the
typical song, maintainer, and owner information. It's kept simple, and
we can add variables as needed.

The second stage is `header.ily` where those values in `info.ily` get mapped
to their special variable counterpart in the `\header` block. This file
is also where one can use the special and standard header information variables
to do custom markup formatting.

And a special circumstance is also the `copyright` and `tagline`
variables. Because we've added standard copyright boilerplate to
lyInclude that uses typical `\header` values, we needed a way to pull that
information without referencing the `\header` block itself. Thus, the
two stages.

Specifying the desired copyright and tagline text is done in the
`header.ily` file and not the `info.ily` file.