feat({layout,settings}.ily): scoreStyle(), part transposing, partChords
`scoreStyle()` is a wrapper function that has been added to settings.ily
that uses tags to filter out music tagged as either "transpose" or
"concert". This change has been added to the layout file as well. Every part is
now made up of 2 duplicate staff halves, one in concert, and the other
transposed. This lets one be able to choose between concert scores and
transposed scores as output without manually tweaking the arguments to
`\transpose`. This also can allow for concert midi playback with
transposed scores in the future as well.

NOTE: that tag filtering is recursive. So if we use
`\keepWithTag`, it will result in any other tags deeper in the music
(fragments for example) getting removed from the end result. So we use
`\removeWithTag` instead to just remove the staff half we dont' want,
keeping the rest.

`partChords` is a tag applied to every set of individual parts chord change
symbols. In situations where the score produces many staves all
containing the same chord changes in the same key (a rhythm section in a
jazz band for example), we want to be able to
remove them from the score, but keep them in for the individual part. My
preference here is to only have one set of chord changes in a score for
staves in concert pitch, but leave them in for instruments that
transpose if they have solo sections. Other combinations are possible
however using this method.